# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Skladi\u0161\u010de API",
    "description": "Inventory management API for tracking physical items and who holds them. All item movements are modeled as transfers between owners (people or locations).",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/users": {
      "get": {
        "summary": "List users",
        "tags": [
          "Users"
        ],
        "description": "Admin only.",
        "responses": {
          "200": {
            "description": "List of users",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/Error"
          }
        }
      },
      "post": {
        "summary": "Create user",
        "tags": [
          "Users"
        ],
        "description": "Admin only. No open registration.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "username",
                  "password",
                  "role"
                ],
                "properties": {
                  "username": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  },
                  "role": {
                    "type": "string",
                    "enum": [
                      "admin",
                      "manager",
                      "user"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "409": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
...
    "/api/transfers": {
      "get": {
        "summary": "List transfers",
        "tags": [
          "Transfers"
        ],
        "description": "All roles. Optionally filter by item or owner.",
        "parameters": [
          {
            "name": "item_id",
            "in": "query",
            "schema": {
              "type": "integer"
            },
            "description": "Filter by item ID"
          },
          {
            "name": "owner_id",
            "in": "query",
            "schema": {
              "type": "integer"
            },
            "description": "Filter by owner ID (matches from or to)"
          }
        ],
        "responses": {
          "200": {
            "description": "List of transfers",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Transfer"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create transfer",
        "tags": [
          "Transfers"
        ],
        "description": "All roles. Moves a quantity of an item from one owner to another. Fails if source doesn't hold enough or if from_owner_id equals to_owner_id.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "item_id",
                  "from_owner_id",
                  "to_owner_id",
                  "quantity"
                ],
                "properties": {
                  "item_id": {
                    "type": "integer",
                    "description": "ID of the item to transfer"
                  },
                  "from_owner_id": {
                    "type": "integer",
                    "description": "Source owner ID"
                  },
                  "to_owner_id": {
                    "type": "integer",
                    "description": "Destination owner ID"
                  },
                  "quantity": {
                    "type": "integer",
                    "minimum": 1,
                    "description": "Number of items to transfer"
                  },
                  "notes": {
                    "type": "string",
                    "description": "Optional notes about the transfer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Transfer created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Transfer"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Get a token from POST /api/auth/login, then pass it as: Authorization: Bearer <token>"
      }
    },
...
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "username": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "admin",
              "manager",
              "user"
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "deleted_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
...
      "Transfer": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "item_id": {
            "type": "integer"
          },
          "from_owner_id": {
            "type": "integer"
          },
          "to_owner_id": {
            "type": "integer"
          },
          "quantity": {
            "type": "integer"
          },
          "notes": {
            "type": "string"
          },
          "transferred_at": {
            "type": "string",
            "format": "date-time"
          },
          "transferred_by": {
            "type": "integer",
            "nullable": true,
            "description": "User ID who performed the transfer"
          },
          "item_name": {
            "type": "string",
            "description": "Joined item name"
          },
          "from_owner_name": {
            "type": "string",
            "description": "Joined source owner name"
          },
          "to_owner_name": {
            "type": "string",
            "description": "Joined destination owner name"
          }
        }
      },
...
    },
    "responses": {
      "Error": {
        "description": "Error response",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "description": "Error message"
                }
              }
            }
          }
        }
      },
...
    }
  }
}